Inventory Navigation + Using Items
Added new controls to the inventory to allow the user to navigate the
action buttons (equip, etc.)

Implemented equiping and using of items. Equipped items must be put away
before another weapon can be equipped and items that can be used can be
used whether something is equipped or not.