Added Enemy Class § Updated Weapon Damage
Added the enemy class that can take damage from weapons (and die when
the health reaches, or gets lower than, 0.

Updated the way weapons handle damage: they now send the value to the
damaged enemy and the enemy will deal with removing health and, when it
comes to destroying, destroying the enemy from the scene.