Item Animations
Added "DefaultAnimation" variable to ARoadFeverCharacterNed; this stores
the Animation Blueprint to use when Ned does not have any equipped
items.

Added "AnimationBlueprint_Default" variable to AItem; this stores the
Animation Blueprint to use when Ned has this item equipped, but is not
aiming with it.

Added "AnimationBlueprint_Aiming" variable to AItem; this stores the
Animation Blueprint to use when Ned has this item equipped, and is
aiming with it.