Weapons Retain Ammo § Reload Base
Made it so items now retain ammo when being equipped and un-equipped.

Added a basic reload (takes ammo out of the gun and, if there is enough
room, puts it in the inventory; if there isn't, it should get thrown in
front of the player).

Added a little UI overlay above the character so one can see how much
ammo is left in the gun (for debugging)