Added Reload Interrupt § Fixed Ammo Bug § UMG Bug
When you are reloading a single round weapon, it will now allow you to
interrupt it by trying to shoot.

There was an issue with ammo where the current ammo affected single
round weapons, however, that should use the stack count; I have switched
this over and it now makes the inventory look a bit cleaner.

There was an issue in the UMG where it wouldn't remove certain items; I
have fixed this (can't remember how, though, so that is good)