Fixed Reload Issue § Fixed Inventory Issue § Fixed Pausing Issue
When switching weapons with a weapon equipped, the weapon wouldn't
reload; this was due to a connection error in Blueprint where I didn't
carry the ammo type over.

The inventory would apear over the top of the pickup confirmation; I
added a check so this no longer happens.

The game now fully pauses when you are in the pickup confirmation menu;
I may add it into the inventory if that is wanted.