Fixed item removal
There was an issue where items would still take up a space in the
inventory unless you open and close the inventory UI; that was due to
the code being inside the slots, rather than the inventory class. I have
moved this code over to the class and it now works fine.