Improved the C++ some more § made collision better.
I re-wrote some of the C++ to get stuff done in less, added more object
checks before using objects.

To accomodate the aforementioned "lessening" of the C++, I had to change
the collisions a bit so that the collision was handled by the engine,
rather than by the C++ (such as the ignored Actors, and what-not).